my-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow==18.1.0
python-dateutil==2.9.0
